[hsmtool] Attribute conversions for dealing with PKCS#11
The `cryptoki` library's type definitions do not support serialization
with serde or conversions to/from their string forms.  Since we want
to be able to define HSM operations and templates as json files, we'll
create a compatible set of PKCS#11 types that support serde, string
conversions and can easily be expressed with json.

1. Autogenerate the various PKCS#11 constant enumerations from the
   cryptoki-sys bindings.
2. Provide conversions for strings and binary data so we can more easily
   represent PKCS#11 binary data (e.g. the public exponent 65537 is
   represented as "01:00:01")
3. Provide conversions between a key-value mapping form and the PKCS#11
   list-of-attributes form.

Signed-off-by: Chris Frantz <cfrantz@google.com>