[bazel] update rules_rust to 0.56
This version splits bindgen to a separate extension, so rename
accordingly.

Updating rules_rust_bindgen pulls in a new version of bindgen which
requires a higher Rust version then what we currently use. So update
Rust nightly to the version that latest stable compiler branches from.
Tock needs an older Rust nightly version to compile, so a new repository_set
is used for it instead.

The new version of clippy starts to warn about needless lifetimes in
cases where it's necessary or idiomatic to keep (e.g. trait
implementations), so disable it for now.

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>